requirements.txt
torch>=1.7.1
torchvision>=0.8.2
pytorch-lightning>=1.4.0
pytorch-geometric>=2.0.0
omegaconf>=2.1.0
numpy>=1.19.2
Pillow>=8.2.0
fvcore>=0.1.5
timm>=0.4.12
einops>=0.3.0
lpips>=0.1.4
wandb>=0.12.0
python-dotenv>=0.19.0
